+++ pkg/ffuf/interfaces.go
@@ -8,14 +8,18 @@ import (
 type MatcherManager interface {
 	SetCalibrated(calibrated bool)
 	SetCalibratedForHost(host string, calibrated bool)
+	SetCalibratedForPath(host string, calibrated bool)
 	AddFilter(name string, option string, replace bool) error
 	AddPerDomainFilter(domain string, name string, option string) error
+	AddPerPathFilter(path string, name string, option string) error
 	RemoveFilter(name string)
 	AddMatcher(name string, option string) error
 	GetFilters() map[string]FilterProvider
 	GetMatchers() map[string]FilterProvider
 	FiltersForDomain(domain string) map[string]FilterProvider
+	FiltersForPath(path string) map[string]FilterProvider
 	CalibratedForDomain(domain string) bool
+	CalibratedForPath(path string) bool
 	Calibrated() bool
 }