+++ pkg/ffuf/job.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"math/rand"
+	"net/url"
 	"os"
 	"os/signal"
 	"sync"
@@ -341,7 +342,10 @@ func (j *Job) isMatch(resp Response) bool {
 	matched := false
 	var matchers map[string]FilterProvider
 	var filters map[string]FilterProvider
-	if j.Config.AutoCalibrationPerHost {
+	if j.Config.AutoCalibrationPerPath {
+		parsed, _ := url.Parse(resp.Request.Url)
+		filters = j.Config.MatcherManager.FiltersForPath(parsed.Path)
+	} else if j.Config.AutoCalibrationPerHost {
 		filters = j.Config.MatcherManager.FiltersForDomain(HostURLFromRequest(*resp.Request))
 	} else {
 		filters = j.Config.MatcherManager.GetFilters()
@@ -443,7 +447,8 @@ func (j *Job) runTask(input map[string][]byte, position int, retried bool) {
 	}
 
 	// Handle autocalibration, must be done after the actual request to ensure sane value in req.Host
-	_ = j.CalibrateIfNeeded(HostURLFromRequest(req), input)
+	parsed, _ := url.Parse(req.Url)
+	_ = j.CalibrateIfNeeded(HostURLFromRequest(req), parsed.Path, input)
 
 	// Handle scraper actions
 	if j.Scraper != nil {
@@ -473,15 +478,16 @@ func (j *Job) runTask(input map[string][]byte, position int, retried bool) {
 		if j.Config.Recursion && j.Config.RecursionStrategy == "greedy" {
 			j.handleGreedyRecursionJob(resp)
 		}
-	} else {
-		if len(resp.ScraperData) > 0 {
-			// print the result anyway, as scraper found something
-			j.Output.Result(resp)
+		if j.Config.Recursion && j.Config.RecursionStrategy == "default" && intSliceContains(j.Config.RecursionStatus, int(resp.StatusCode)) {
+			j.handleDefaultRecursionJob(resp)
 		}
+
+		return
 	}
 
-	if j.Config.Recursion && j.Config.RecursionStrategy == "default" && len(resp.GetRedirectLocation(false)) > 0 {
-		j.handleDefaultRecursionJob(resp)
+	if len(resp.ScraperData) > 0 {
+		// print the result anyway, as scraper found something
+		j.Output.Result(resp)
 	}
 }
 
@@ -511,7 +517,7 @@ func (j *Job) handleGreedyRecursionJob(resp Response) {
 // not been reached
 func (j *Job) handleDefaultRecursionJob(resp Response) {
 	recUrl := resp.Request.Url + "/" + "FUZZ"
-	if (resp.Request.Url + "/") != resp.GetRedirectLocation(true) {
+	if (resp.Request.Url+"/") != resp.GetRedirectLocation(true) && resp.StatusCode >= 300 && resp.StatusCode < 400 {
 		// Not a directory, return early
 		return
 	}
@@ -593,4 +599,13 @@ func (j *Job) Stop() {
 // Stop current, resume to next
 func (j *Job) Next() {
 	j.RunningJob = false
+}
+
+func intSliceContains(sslice []int, num int) bool {
+	for _, v := range sslice {
+		if v == num {
+			return true
+		}
+	}
+	return false
 }