+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   - New
     - New autocalibration strategy "extra" that uses a list of commonly blacklisted files to filter blacklist responses
     - New command line parameter for auto adjasting ratelimiting based on amount of 429 `-ar`
+    - New command line parameter for recursive autocalibration on every path `-acp`
+    - New command line parameter for controlling recursion status code `-recursion-status`
   - Changed
     - Explicitly allow TLS1.0 
     - Fix markdown output file format